Component becomes LayoutDOM
- Plot, Widget, and Layouts inherit from LayoutDOM
- LayoutDOM has width, height, and disabled properties
- Remove width and height from Chart (were deprecated anyway)
