+++ bokeh/models/plots.py
@@ -24,7 +24,8 @@ from .ranges import Range, Range1d, FactorRange
 from .renderers import Renderer, GlyphRenderer, DataRenderer, TileRenderer, DynamicImageRenderer
 from .sources import DataSource, ColumnDataSource
 from .tools import Tool, ToolEvents
-from .component import Component
+from .layouts import LayoutDOM
+
 
 def _select_helper(args, kwargs):
     """
@@ -56,12 +57,14 @@ def _select_helper(args, kwargs):
         selector = kwargs
     return selector
 
+
 class LayoutBox(Model):
     ''' Represents an **on-canvas** layout.
 
     '''
 
-class Plot(Component):
+
+class Plot(LayoutDOM):
     """ Model representing a plot, containing glyphs, guides, annotations.
 
     """
@@ -611,15 +614,16 @@ class Plot(Component):
     """)
 
 
-class GridPlot(Component):
+
+class GridPlot(LayoutDOM):
     """ A 2D grid of plots rendered on separate canvases in an HTML table.
 
     """
 
-    # TODO (bev) really, GridPlot should be a layout, not a Plot subclass
     @validation.error(REQUIRED_RANGE)
     def _check_required_range(self):
         pass
+
     @validation.warning(MISSING_RENDERERS)
     def _check_missing_renderers(self):
         pass
